RSAKeyGenerator.java
import java.math.BigInteger;
import java.security.SecureRandom;

public class RSAKeyGenerator {
    private static final int BIT_LENGTH = 2048; // Adjust the bit length as needed

    public static void main(String[] args) {
        // Example usage to generate keys for a user
        String userId = "alice";
        KeyPair keyPair = generateKeyPair(userId);

        // Print the generated keys
        System.out.println("Public Key: " + keyPair.getPublicKey());
        System.out.println("Private Key: " + keyPair.getPrivateKey());
    }

    public static KeyPair generateKeyPair(String userId) {
        // Step 1: Choose two large prime numbers, p and q
        BigInteger p = generateLargePrime();
        BigInteger q = generateLargePrime();

        // Step 2: Calculate n = p * q
        BigInteger n = p.multiply(q);

        // Step 3: Calculate the totient function φ(n)
        BigInteger phiN = p.subtract(BigInteger.ONE).multiply(q.subtract(BigInteger.ONE));

        // Step 4: Choose public exponent e such that 1 < e < φ(n) and gcd(e, φ(n)) = 1
        BigInteger e = choosePublicExponent(phiN);

        // Step 5: Calculate private exponent d such that (d * e) % φ(n) = 1
        BigInteger d = calculatePrivateExponent(e, phiN);

        // Create public and private keys
        PublicKey publicKey = new PublicKey(n, e);
        PrivateKey privateKey = new PrivateKey(n, d);

        // Associate keys with the user ID
        KeyPair keyPair = new KeyPair(userId, publicKey, privateKey);

        return keyPair;
    }

    private static BigInteger generateLargePrime() {
        // Implement a method to generate a large prime number
        // You can use the BigInteger.probablePrime() method or a custom method
        return BigInteger.probablePrime(BIT_LENGTH, new SecureRandom());
    }

    private static BigInteger choosePublicExponent(BigInteger phiN) {
        // Implement a method to choose a public exponent 'e'
        // Common choice is 65537, but you can implement a more sophisticated method
        return BigInteger.valueOf(65537);
    }

    private static BigInteger calculatePrivateExponent(BigInteger e, BigInteger phiN) {
        // Implement a method to calculate the private exponent 'd'
        return e.modInverse(phiN);
    }
}
